src/main/webapp/app/entities/configuration-salon/delete/configuration-salon-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ConfigurationSalonService } from '../service/configuration-salon.service';

import { ConfigurationSalonDeleteDialogComponent } from './configuration-salon-delete-dialog.component';

describe('ConfigurationSalon Management Delete Component', () => {
  let comp: ConfigurationSalonDeleteDialogComponent;
  let fixture: ComponentFixture<ConfigurationSalonDeleteDialogComponent>;
  let service: ConfigurationSalonService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [ConfigurationSalonDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(ConfigurationSalonDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ConfigurationSalonDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ConfigurationSalonService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('9fec3727-3421-4967-b213-ba36557ca194');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
